+++ src/components/pages/ObserveBudget.js
@@ -5,12 +5,12 @@ import GetBudget from '../GetBudgets'
 export default class ObserveBudget extends Component {
     render() {
         return (
-        <div>
+            <div>
                 <Navbar />
-                <p>hej från observe budget</p>
+
 
                 <GetBudget />
-        </div>
+            </div>
         )
     }
 }
